refactor: modularize utils.py, extracting cli and chapter detector modules
